0.8.4.5: 	Make simple-streams work on alpha 	... really ensure the various ffi_foo_t types for wrapped_stat 		are 32-bit everywhere 	... update the lisp version of the struct definition to believe 		this, too (UNSIGNED-INT, not UNSIGNED-LONG) 	... also bugfix: MOST-POSITIVE-FIXNUM is an upper exclusive bound 	... stray arch_os_get_current_thread() fix

+++ src/runtime/wrap.c
@@ -150,8 +150,11 @@ wrapped_readlink(char *path)
  * when values are small.
  *
  * FIXME: But of course we should fix the FFI so that we can use the
- * actual 64-bit values instead. */
-typedef long ffi_dev_t; /* since Linux dev_t can be 64 bits */
+ * actual 64-bit values instead.  In fact, we probably have by now
+ * (2003-10-03) on all working platforms except MIPS and HPPA; if some
+ * motivated spark would simply fix those, this hack could go away.
+ * -- CSR, 2003-10-03 */
+typedef u32 ffi_dev_t; /* since Linux dev_t can be 64 bits */
 typedef u32 ffi_off_t; /* since OpenBSD 2.8 st_size is 64 bits */
 
 /* a representation of stat(2) results which doesn't depend on CPU or OS */
